* bin/pitivi.in: * pitivi/bin.py: * pitivi/objectfactory.py: * pitivi/settings.py: * pitivi/threads.py: * pitivi/thumbnailer.py: * pitivi/timeline/composition.py: * pitivi/timeline/objects.py: * pitivi/timeline/source.py: * pitivi/timeline/timeline.py: * pitivi/ui/mainwindow.py: * pitivi/ui/ruler.py: * pitivi/ui/slider.py: * pitivi/ui/sourcefactories.py: * pitivi/ui/timeline.py: * pitivi/ui/timelineobjects.py: * pitivi/ui/tracklayer.py: * pitivi/ui/viewer.py: Fix misc issues detected by pychecker. Renaming unused method arguments Whitespace/tabs fixes.
git-svn-id: svn+ssh://svn.gnome.org/svn/pitivi/trunk@1099 d3729300-e425-0410-8a4c-d956edccc248

+++ pitivi/ui/ruler.py
@@ -249,7 +249,7 @@ class ScaleRuler(gtk.Layout, ZoomableWidgetInterface):
 
             if timeprint == 0:
                 # draw the text position
-                hours = seconds / 3600
+                hours = int(seconds / 3600)
                 mins = seconds % 3600 / 60
                 secs = seconds % 60
                 time = "%02d:%02d:%02d" % (hours, mins, secs)
